Handle transaction when CA not needed
